pyproject.toml
[project]
name = "agent-scheduler"
version = "0.1.0"
description = "Scheduled execution of AI coding agent CLIs"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "rich>=13",
    "pydantic>=2",
    "pydantic-settings>=2",
    "python-crontab>=3",
    "python-dotenv>=1",
]

[project.scripts]
agent-scheduler = "agent_scheduler.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_scheduler"]

[tool.pytest.ini_options]
testpaths = ["tests"]
